requirements.txt
pandas==2.1.0
neo4j==5.14.1
tqdm==4.66.1
python-dotenv==1.0.0 